Refactor raw parse package, Change translation strategy
Refactored the raw parse package to new layout.

Changed the expasion of raw objects to be more self contained.
Expanded objects to represent maximum data to be had from bid package.
Ment to be a stable interface that can be changed into other formats to suit.
Not ment to be serializable at this time.